ZipChemCompProvider cleanup: * Moved to use Java 7 ZipFileSystem for appending files and refactored * ZipAppendUtil code to be within ZipChemCompProvider. * Cleaned up exception handling within ZipChemCompProvider * Created optional mode for ZipChemCompProvider that doesn't delete  .cif.gz - this prevents redownloading .cif.gz with unit testing. * ZipChemCompProvider is thread-safe, with reads/appends in synchronized  methods.

+++ biojava-structure/src/main/java/org/biojava/nbio/structure/io/mmcif/ZipChemCompProvider.java
@@ -21,9 +21,11 @@
 package org.biojava.nbio.structure.io.mmcif;
 
 import java.io.*;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.net.URI;
+import java.nio.file.*;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 import java.util.logging.Logger;
 import java.util.zip.GZIPInputStream;
@@ -36,14 +38,17 @@ import org.biojava.nbio.structure.io.mmcif.chem.PolymerType;
 import org.biojava.nbio.structure.io.mmcif.chem.ResidueType;
 import org.biojava.nbio.structure.io.mmcif.model.ChemComp;
 
-/** This chemical component provider retrieves and caches chemical component definition files from a zip archive specified in its construction.
- * 	If the archive does not contain the record, an attempt is made to download it using DownloadChemCompProvider. The downloaded file is then added to the archive.
- *	It is up to the calling function to handle exceptions raised by constructing with a non-existent archive. 
+/** This chemical component provider retrieves and caches chemical component definition files from a
+ * zip archive specified in its construction.  If the archive does not contain the record, an attempt is
+ * made to download it using DownloadChemCompProvider. The downloaded file is then added to the archive.
+ *
+ * The class is thread-safe and the same ZipChemCompProvider should be used by all threads to prevent
+ * simultaneous read or write to the zip archive.  A zip archive will be created if missing.
  *
  * @author edlunde
  * @author larsonm
  * @since 12/05/12
- * 
+ * updated 3/5/2016 for Java 7 ZipFileSystem
  */
 public class ZipChemCompProvider implements ChemCompProvider{
 	private static final Logger s_logger = Logger.getLogger( ZipChemCompProvider.class.getPackage().getName() );
@@ -52,21 +57,21 @@ public class ZipChemCompProvider implements ChemCompProvider{
 	private final Path m_zipRootDir;
 	private final Path m_zipFile;
 	private final DownloadChemCompProvider m_dlProvider;
-	// private final ZipFile m_zipDictionary;
-	
+
+	private boolean m_removeCif;
+
 	// Missing IDs from library that cannot be download added here to prevent delays.
 	private Set<String> unavailable = new HashSet<String>();
 
 	/**
-	 * Constructor for a ZipChemCompProvider, a Chemical Component provider that stores chemical component
-	 * files in a single zip archive.  Missing chemical components will be downloaded and appended to this
-	 * archive.
+	 * ZipChemCompProvider is a Chemical Component provider that stores chemical components
+	 * in a zip archive.  Missing chemical components are downloaded and appended to the
+	 * archive.  If non-existent a new zip archive will be created.
 	 *
-	 * @param chemicalComponentDictionaryFile : path to zip archive for chemical components. If non-existent,
-	 *                                        an empty zip archive will be created and populated.
-	 * @param tempDir : path for temporary directory, (null) will system default property "java.io.tmpdir".
+	 * @param chemicalComponentDictionaryFile : path to zip archive for chemical components.
+	 * @param tempDir : path for temporary directory, (null) defaults to path in property "java.io.tmpdir".
 	 * @throws IOException
-	 */
+     */
 	public ZipChemCompProvider(String chemicalComponentDictionaryFile, String tempDir) throws IOException {
 		this.m_zipFile = Paths.get(chemicalComponentDictionaryFile);
 
@@ -81,20 +86,18 @@ public class ZipChemCompProvider implements ChemCompProvider{
 
 		// Setup an instance of the download chemcomp provider.
 		this.m_dlProvider = new DownloadChemCompProvider(m_tempDir.toString());
-
-		// Create a new zip dictionary file if one isn't already created.
-		initializeZipFile();
+		initializeZip();
 	}
-	
-	// Create the zip file and setup a fallback DownloadChemCompProvider.
-	private void initializeZipFile() throws IOException {
-		s_logger.info("Using chemical component dictionary: " + m_zipFile.toString());
 
-		// Check if file exists, if not create.
-		final File zipFile = m_zipFile.toFile();
-		if (!zipFile.exists()) {
-			FileOutputStream f = new FileOutputStream(zipFile);
-			ZipOutputStream zip = new ZipOutputStream(new BufferedOutputStream(f));
+	// See comments in addToZipFileSystem for why initialization is required with
+	// ZipFileSystems - due to URI issues in Java7.
+	private void initializeZip() throws IOException {
+		s_logger.info("Using chemical component dictionary: " + m_zipFile.toString());
+		final File f = m_zipFile.toFile();
+		if (!f.exists()) {
+			s_logger.info("Creating missing zip archive: " + m_zipFile.toString());
+			FileOutputStream fo = new FileOutputStream(f);
+			ZipOutputStream zip = new ZipOutputStream(new BufferedOutputStream(fo));
 			try {
 				zip.putNextEntry(new ZipEntry("chemcomp/"));
 				zip.closeEntry();
@@ -102,6 +105,15 @@ public class ZipChemCompProvider implements ChemCompProvider{
 				zip.close();
 			}
 		}
+	}
+
+	/**
+	 * Remove downloaded .cif.gz after adding to zip archive?
+	 * Default is true.
+	 * @param doRemove
+     */
+	public void setRemoveCif(boolean doRemove) {
+		m_removeCif = doRemove;
 	}
 	
 	/* (non-Javadoc)
@@ -119,40 +131,15 @@ public class ZipChemCompProvider implements ChemCompProvider{
 			if (recordName.equals(str)) return getEmptyChemComp(recordName);
 		}
 
-		ChemComp cc = null; // Will find in zip or download
-		String filename = "chemcomp/" + recordName+".cif.gz";
-
-		// Try to pull from zip.
-		try {
-			ZipFile m_zip = new ZipFile(m_zipFile.toFile());
-			ZipEntry ccEntry = m_zip.getEntry(filename);
-			if (ccEntry != null) {
-				final InputStream zipStream = m_zip.getInputStream(ccEntry);
-				final InputStream inputStream = new GZIPInputStream(zipStream);
-				s_logger.fine("reading " + recordName + " from " + m_zipFile);
-				final MMcifParser parser = new SimpleMMcifParser();
-				final ChemCompConsumer consumer = new ChemCompConsumer();
-				parser.addMMcifConsumer(consumer);
-				parser.parse(inputStream);
-
-				final ChemicalComponentDictionary dict = consumer.getDictionary();
-				cc = dict.getChemComp(recordName);
-				inputStream.close();
-			}
-			m_zip.close();
-		} catch (IOException ex) {
-			s_logger.severe("Error reading " + m_zipFile.toString() + " " + ex.getMessage());
-		}
-
-		// Try to download next.
+		// Try to pull from zip, if fail then download.
+		ChemComp cc = getFromZip(recordName);
 		if (cc == null) {
 			s_logger.info("File "+recordName+" not found in archive. Attempting download from PDB.");
 			cc = downloadAndAdd(recordName);
 		}
 
-		// Still failed, mark unavailable and return empty chemcomp
-		if (cc == null) {
-			// Could not find this ChemComp, add this to a list of unavailable.
+		// If a null record or an empty chemcomp, return a default ChemComp and blacklist.
+		if (cc == null || (null == cc.getName() && cc.getAtoms().size() == 0)) {
 			s_logger.info("Unable to find or download " + recordName + " - excluding from future searches.");
 			unavailable.add(recordName);
 			return getEmptyChemComp(recordName);
@@ -168,37 +155,30 @@ public class ZipChemCompProvider implements ChemCompProvider{
      */
 	private ChemComp downloadAndAdd(String recordName){
 		final ChemComp cc = m_dlProvider.getChemComp(recordName);
-		
-		if (cc != null && "?".equals(cc.getOne_letter_code())) {
-			// Failed to find this chemical component and defaulted back to reduced chemical component.
-			unavailable.add(recordName);
-			return cc;  // Don't try to add this to the chemcomp.zip.
-		}
 
-		//DownloadProvider places files in default temp dir\chemcomp
-		final File [] files = finder(m_tempDir.resolve("chemcomp").toString(), "cif.gz");
+		// final File [] files = finder(m_tempDir.resolve("chemcomp").toString(), "cif.gz");
+		final File [] files = new File[1];
+		Path cif = m_tempDir.resolve("chemcomp").resolve(recordName + ".cif.gz");
+		files[0] = cif.toFile();
 		if (files != null) {
-			ZipAppendUtil.getInstance().addToZipFileSystem(m_zipFile, files, m_zipRootDir);
-			for (File f : files) f.delete();
+			addToZipFileSystem(m_zipFile, files, m_zipRootDir);
+			if (m_removeCif) for (File f : files) f.delete();
 		}
 		return cc;
 	}
 
 	/**
-	 * Cleanup any temporary chemical component files that have been created within tmpdir.
-	 * @param tempdir : path to temporary directory for chemical component downloads.
+	 * Cleanup chemical component (.cif.gz) files downloaded to tmpdir.
+	 * @param tempdir : path to temporary directory for chemical components
      */
-	public static void purgeAllTempFiles(String tempdir) {
-		if (null != tempdir) {
-			File[] chemCompOutFiles = finder(tempdir, "cif");
-			for (File f : chemCompOutFiles) f.delete();
+	public static void purgeTempFiles(String tempdir) {
+		if (tempdir == null) return;
 
-			File[] chemCompTempFiles = finderPrefix(tempdir, "chemcomp.zip");
-			for (File f : chemCompTempFiles) f.delete();
-		}
-
-		Path chemCompDirPath = Paths.get(System.getProperty(tempdir)).resolve("chemcomp");
-		chemCompDirPath.toFile().delete();
+		s_logger.info("Removing: "+tempdir);
+		Path dlPath = Paths.get(tempdir).resolve("chemcomp");
+		File[] chemCompOutFiles = finder(dlPath.toString(), "cif.gz");
+		if (null != chemCompOutFiles) for (File f : chemCompOutFiles) f.delete();
+		dlPath.toFile().delete();
 	}
 
 	/**
@@ -236,18 +216,90 @@ public class ZipChemCompProvider implements ChemCompProvider{
 			{ return filename.endsWith(suffix); }
 		} );
 	}
-	
+
 	/**
-	 * Return File(s) in dirName that match prefix.
-	 * @param dirName
-	 * @param prefix
-	 * @return
+	 * This is synchronized, along with addToFileSystem to prevent simulatenous reading/writing.
+	 * @param recordName to find in zipfile.
+	 * @return ChemComp if found or null if missing.
+     */
+	private synchronized ChemComp getFromZip(String recordName) {
+		ChemComp cc = null;
+		if (!m_zipFile.toFile().exists()) return cc;
+		final String filename = "chemcomp/" + recordName+".cif.gz";
+
+		// try with resources block to read from the filesystem.
+		try (FileSystem fs = FileSystems.newFileSystem(m_zipFile, null)) {
+			Path cif = fs.getPath(filename);
+
+			if (Files.exists(cif)) {
+				final InputStream zipStream = Files.newInputStream(cif);
+				final InputStream inputStream = new GZIPInputStream(zipStream);
+				s_logger.fine("reading " + recordName + " from " + m_zipFile);
+				final MMcifParser parser = new SimpleMMcifParser();
+				final ChemCompConsumer consumer = new ChemCompConsumer();
+				parser.addMMcifConsumer(consumer);
+				parser.parse(inputStream);
+				inputStream.close();
+
+				final ChemicalComponentDictionary dict = consumer.getDictionary();
+				cc = dict.getChemComp(recordName);
+			}
+		} catch (IOException e) {
+			s_logger.severe("Unable to read from zip file : " + e.getMessage());
+		}
+
+		return cc;
+	}
+
+	/**
+	 * Add an array of files to a zip archive.
+	 * Synchronized to prevent simultaneous reading/writing.
+	 *
+	 * @param zipFile is a destination zip archive
+	 * @param files is an array of files to be added
+	 * @param pathWithinArchive is the path within the archive to add files to
+	 * @return true if successfully appended these files.
 	 */
-	static private File[] finderPrefix( String dirName, final String prefix){
-		final File dir = new File(dirName);
-		return dir.listFiles(new FilenameFilter() { 
-			public boolean accept(File dir, String filename)
-			{ return filename.startsWith(prefix); }
-		} );
+	private synchronized boolean addToZipFileSystem(Path zipFile, File[] files, Path pathWithinArchive) {
+		boolean ret = false;
+
+		/* URIs in Java 7 cannot have spaces, must use Path instead
+		 * and so, cannot use the properties map to describe need to create
+		 * a new zip archive.  ZipChemCompProvider.initilizeZip to creates the
+		 * missing zip file */
+
+		/*
+		// convert the filename to a URI
+		String uriString = "jar:file:" + zipFile.toUri().getPath();
+		final URI uri = URI.create(uriString);
+
+		// if filesystem doesn't exist, create one.
+		final Map<String, String> env = new HashMap<>();
+		// Create a new zip if one isn't present.
+		if (!zipFile.toFile().exists()) {
+			System.out.println("Need to create " + zipFile.toString());
+		}
+		env.put("create", String.valueOf(!zipFile.toFile().exists()));
+		// Specify the encoding as UTF -8
+		env.put("encoding", "UTF-8");
+		*/
+
+		// Copy in each file.
+		try (FileSystem zipfs = FileSystems.newFileSystem(zipFile, null)) {
+			Files.createDirectories(pathWithinArchive);
+			for (File f : files) {
+				if (!f.isDirectory() && f.exists()) {
+					Path externalFile = f.toPath();
+					Path pathInZipFile = zipfs.getPath(pathWithinArchive.resolve(f.getName()).toString());
+					Files.copy(externalFile, pathInZipFile,
+							StandardCopyOption.REPLACE_EXISTING);
+				}
+			}
+			ret = true;
+		} catch (IOException ex) {
+			s_logger.severe("Unable to add entries to Chemical Component zip archive : " + ex.getMessage());
+			ret = false;
+		}
+		return ret;
 	}
 }
